Modify double to int64_t

+++ Eliptic_curve_secret_sharing/include/recov.h
@@ -2,6 +2,6 @@
 #define _RECOV_H_
 
 void interpolation(uint64_t m, BIGNUM ***point_p, BIGNUM *secret);
-double recovery(uint64_t m, EC_GROUP *curve, EC_POINT *generator, BIGNUM *secret);
+int64_t recovery(uint64_t m, EC_GROUP *curve, EC_POINT *generator, BIGNUM *secret);
 
 #endif
